+++ example/app/main.js
@@ -6,10 +6,13 @@ const createWindow = () => {
   win = new BrowserWindow({
     width: 800,
     minWidth: 500,
-    height: 620,
-    minHeight: 500,
+    height: 520,
+    minHeight: 400,
     center: true,
     show: false,
+    webPreferences: {
+      nodeIntegration: true,
+    },
   })
   win.loadURL(`file://${__dirname}/index.html`)