+++ src/apis/Alert.js
@@ -9,25 +9,28 @@ type AlertButton = {
 
 type AlertType = 'none' | 'info' | 'error' | 'question' | 'warning'
 
+type AlertRes = {
+  response: number,
+  checkboxChecked: boolean,
+}
+
 export const alert = (
   title: string,
   message: ?string,
   buttons: AlertButton[] = [],
   type: AlertType = 'none',
 ) => {
-  remote.dialog.showMessageBox(
-    remote.getCurrentWindow(),
-    {
+  remote.dialog
+    .showMessageBox(remote.getCurrentWindow(), {
       type,
       buttons: buttons.map(b => b.text),
       message: title,
       detail: message,
-    },
-    (index: number) => {
-      const button = buttons[index]
+    })
+    .then(({ response }: AlertRes) => {
+      const button = buttons[response]
       if (button && button.onPress) {
         button.onPress()
       }
-    },
-  )
+    })
 }